[ir] Add loading test for IR
We didn't really test all the input functionality and found a couple of
bugs (like actually not loading triggers).

We added a test that loads a specification from a yaml and ensures that
we process the same thing.

This should safe time in the short term as the backend can in general
assume things are correct.